pushing changes before pulling down
