+++ client/stylesheets/ChatFire.css
@@ -13,6 +13,7 @@ h1 {
   height: 400px;
   overflow: scroll;
   z-index: 100;
+  border: 0.5px double grey;
 }
 
 .msg-btn {
