+++ client/stylesheets/CollabBox.css
@@ -9,6 +9,7 @@
   width: 100%;
   background-color: rgba(0, 0, 0, 0.2);
   z-index: 100;
+  overflow: scroll;
 }
 
 .chat-box {
